change solution after comments
